+++ tests/mocks/data/index.ts
@@ -1,3 +1,4 @@
 export * from "./mock-account";
+export * from "./mock-assets";
 export * from "./mock-portfolio";
 export * from "./mock-prisma";
