+++ tests/mocks/repositories/account-repository-spy.ts
@@ -2,7 +2,10 @@ import { faker } from "@faker-js/faker";
 
 import { Account, OAuthUser } from "@/domain/models";
 
-import { AddAccountRepository } from "@/domain/contracts/repositories";
+import {
+  AddAccountRepository,
+  LoadAccountByEmailRepository,
+} from "@/domain/contracts/repositories";
 
 export class AddAccountRepositorySpy implements AddAccountRepository {
   input: OAuthUser = {
@@ -19,6 +22,27 @@ export class AddAccountRepositorySpy implements AddAccountRepository {
   };
 
   async create(input: OAuthUser): Promise<Account> {
+    this.input = input;
+
+    return this.output;
+  }
+}
+
+export class LoadAccountByEmailRepositorySpy
+  implements LoadAccountByEmailRepository
+{
+  input = faker.internet.email();
+
+  output: Account = {
+    id: faker.string.uuid(),
+    name: faker.person.fullName(),
+    email: faker.internet.email(),
+    image: faker.image.avatar(),
+  };
+
+  async loadByEmail(
+    input: string
+  ): Promise<LoadAccountByEmailRepository.Output> {
     this.input = input;
 
     return this.output;
